requirements.txt
# Core dependencies
streamlit>=1.28.0
openai>=1.3.0
python-dotenv>=1.0.0

# LangChain components
langchain>=0.0.350
langchain-openai>=0.0.2
langchain-community>=0.0.10

# Document processing
PyPDF2>=3.0.1
python-docx>=0.8.11

# Vector database
chromadb>=0.4.15

# Text processing
tiktoken>=0.5.1

# HTTP requests (for OpenAI)
httpx>=0.25.0

# Production deployment
gunicorn>=21.2.0